FIX FATAL BUG 😭
When iterating each choice of each question in order to grade the quiz,
the code was `question.querySelectorAll("li")`.
This is a mistake because "li" can also be a list item in the explanation.

The bug only shows up recently because
only recently did people start writing explanations containing lists.

Fix: `question.querySelectorAll(".question-choices li")`